requirements.txt
asgiref==3.7.2
attrs==23.1.0
black==23.3.0
boto3==1.26.153
botocore==1.29.153
certifi==2023.5.7
charset-normalizer==3.1.0
click==8.1.3
colorama==0.4.6
coverage==7.3.0
Django==4.2.2
django-admin-honeypot-updated-2021==1.2.0
django-admin-thumbnails==0.2.8
django-ipware==5.0.0
django-phonenumber-field==7.1.0
django-storages==1.13.2
djangorestframework==3.14.0
djangorestframework-simplejwt==5.3.0
drf-spectacular==0.26.4
factory-boy==3.3.0
Faker==19.3.0
idna==3.4
inflection==0.5.1
iniconfig==2.0.0
jmespath==1.0.1
jsonschema==4.19.0
jsonschema-specifications==2023.7.1
mypy-extensions==1.0.0
packaging==23.1
pathspec==0.11.1
phonenumbers==8.13.18
phonenumberslite==8.13.13
Pillow==9.5.0
platformdirs==3.6.0
pluggy==1.2.0
psycopg2-binary==2.9.6
PyJWT==2.8.0
pytest==7.4.0
pytest-cov==4.1.0
pytest-django==4.5.2
pytest-factoryboy==2.5.1
python-dateutil==2.8.2
python-decouple==3.8
pytz==2023.3
PyYAML==6.0.1
referencing==0.30.2
requests==2.31.0
rpds-py==0.9.2
s3transfer==0.6.1
six==1.16.0
sqlparse==0.4.4
typing_extensions==4.7.1
tzdata==2023.3
uritemplate==4.1.1
urllib3==1.26.16
